Allow the usage of model instance for callable return_instead
With this commit, callables passed to return_instead can accept a single
argument which will be the model instance on which the deprecated field
is being accessed.

This allows users to set the return value of the deprecated field to
that of another field directly (aliasing) or to extrapolate what would
have been the value based on one or more model instance attributes.

This is a breaking change and any functions/methods passed to
return_instead should allow at least one argument.