Handle descriptor access through Model class
This commit makes necessary changes to the DeprecatedField descriptor so
that it behaves as expected when accessed without a Model instance.

The expected behavior when accessed through a Model instance is to
return the value computed / stored by the descriptor.

The expected behavior when accessed through a Model class is to return
the descriptor itself.

This commit also introduces a first test and the necessary libraries and
requirements files to setup said test.